Tests/LibWeb/Text/expected/select-text.txt

Lorem
Lorem ip
dolor
dolor s
dolor sit
First paragraph for triple-click test.
    Second paragraph with a link inside.
